feat: move CardWidget implementation to a separate module for better organization
